udp_packet_challenge/requirements.txt
pycryptodome==3.9.4
locked-dict==0.1.1
